incomplete Library functions

+++ src/main/java/com/irms_service/repository/AddressRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.irms_service.entity.AddressEntity;
 
 @Repository
-public interface AddressRepository extends JpaRepository<AddressEntity, Integer>{
+public interface AddressRepository extends JpaRepository<AddressEntity, Long>{
 
 	
 }
